+++ app/Http/Controllers/IndustryTypeController.php
@@ -4,14 +4,28 @@ namespace App\Http\Controllers;
 
 use App\Models\IndustryType;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class IndustryTypeController extends Controller
 {
+    public $user;
+
+
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = Auth::guard('web')->user();
+            return $next($request);
+        });
+    }
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        if (is_null($this->user) || !$this->user->can('industry-type.index')) {
+            abort(403, 'Unauthorized');
+        }
         $datas = IndustryType::orderBy('industry_seqno')->get();
         return view('admin.industryType.index', compact('datas'));
     }
@@ -29,6 +43,9 @@ class IndustryTypeController extends Controller
      */
     public function store(Request $request)
     {
+        if (is_null($this->user) || !$this->user->can('industry-type.store')) {
+            abort(403, 'Unauthorized');
+        }
         $request->validate([
             'industry_code' => "required|unique:industry_types",
             'industry_desc' => 'required',
@@ -60,6 +77,9 @@ class IndustryTypeController extends Controller
      */
     public function update(Request $request, IndustryType $industry_type)
     {
+        if (is_null($this->user) || !$this->user->can('industry-type.update')) {
+            abort(403, 'Unauthorized');
+        }
         $request->validate([
             'industry_code' => "required|unique:industry_types,industry_code,{$industry_type->id}",
             'industry_desc' => 'required',
@@ -75,6 +95,9 @@ class IndustryTypeController extends Controller
      */
     public function destroy(IndustryType $industry_type)
     {
+        if (is_null($this->user) || !$this->user->can('industry-type.destroy')) {
+            abort(403, 'Unauthorized');
+        }
         $industry_type->delete();
         return back()->with('success', 'Successfully deleted.');
     }
