+++ app/Http/Controllers/ReligionController.php
@@ -4,14 +4,28 @@ namespace App\Http\Controllers;
 
 use App\Models\Religion;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ReligionController extends Controller
 {
+    public $user;
+
+
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = Auth::guard('web')->user();
+            return $next($request);
+        });
+    }
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        if (is_null($this->user) || !$this->user->can('religion.index')) {
+            abort(403, 'Unauthorized');
+        }
         $datas = Religion::latest()->get();
         return view('admin.religion.index', compact('datas'));
     }
@@ -29,6 +43,9 @@ class ReligionController extends Controller
      */
     public function store(Request $request)
     {
+        if (is_null($this->user) || !$this->user->can('religion.store')) {
+            abort(403, 'Unauthorized');
+        }
         $request->validate([
             'religion_code' => 'required|unique:religions',
             'religion_desc' => 'required',
@@ -60,6 +77,9 @@ class ReligionController extends Controller
      */
     public function update(Request $request, Religion $religion)
     {
+        if (is_null($this->user) || !$this->user->can('religion.update')) {
+            abort(403, 'Unauthorized');
+        }
         $request->validate([
             'religion_code' => "required|unique:religions,religion_code,{$religion->id}",
             'religion_desc' => 'required',
@@ -75,6 +95,9 @@ class ReligionController extends Controller
      */
     public function destroy(Religion $religion)
     {
+        if (is_null($this->user) || !$this->user->can('religion.destroy')) {
+            abort(403, 'Unauthorized');
+        }
         $religion->delete();
 
         return back()->with('success', 'Deleted Successfully.');
